checker/pyproject.toml
[tool.poetry]
name = "reference"
version = "0.1.0"
description = ""
authors = ["inqb <inqb@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
requests = "^2.25.1"
PyYAML = "^5.4.1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
